Create a light theme.
This CL moves the current themes/dark to themes/basic. It then creates
a themes/dark and themes/light which both use themes/basic and implement the
CreateTheme method to set the desired theme colours.

The colours used in the light theme are the colours from:
https://github.com/google/gxui/pull/129.

The linear_layout test has been modified to have a -theme argument. It
defaults to dark but dark or light can be passed to set the theme.

BUG=#128
