+++ samples/linear_layout/main.go
@@ -5,15 +5,26 @@
 package main
 
 import (
+	"flag"
 	"github.com/google/gxui"
 	"github.com/google/gxui/drivers/gl"
 	"github.com/google/gxui/math"
-	"github.com/google/gxui/samples/flags"
 	"github.com/google/gxui/themes/dark"
+	"github.com/google/gxui/themes/light"
+)
+
+var (
+	flagTheme          = flag.String("theme", "dark", "Theme to use {dark|light}.")
+	defaultScaleFactor = flag.Float64("scaling", 1.0, "Adjusts the scaling of UI rendering")
 )
 
 func appMain(driver gxui.Driver) {
-	theme := dark.CreateTheme(driver)
+	var theme gxui.Theme
+	if *flagTheme == "light" {
+		theme = light.CreateTheme(driver)
+	} else {
+		theme = dark.CreateTheme(driver)
+	}
 
 	layout := theme.CreateLinearLayout()
 	layout.SetSizeMode(gxui.Fill)
@@ -80,12 +91,13 @@ func appMain(driver gxui.Driver) {
 	update()
 
 	window := theme.CreateWindow(800, 600, "Linear layout")
-	window.SetScale(flags.DefaultScaleFactor)
+	window.SetScale(float32(*defaultScaleFactor))
 	window.AddChild(layout)
 	window.OnClose(driver.Terminate)
 	window.SetPadding(math.Spacing{L: 10, T: 10, R: 10, B: 10})
 }
 
 func main() {
+	flag.Parse()
 	gl.StartDriver(appMain)
 }
