+++ themes/basic/image.go
@@ -2,10 +2,15 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package dark
+package basic
 
 import (
 	"github.com/google/gxui"
+	"github.com/google/gxui/mixins"
 )
 
-var Blue30 = gxui.Color{R: 0.0, G: 0.0, B: 0.3, A: 1.0}
+func CreateImage(theme *Theme) gxui.Image {
+	i := &mixins.Image{}
+	i.Init(i, theme)
+	return i
+}
