+++ themes/basic/linear_layout.go
@@ -2,17 +2,15 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package dark
+package basic
 
 import (
 	"github.com/google/gxui"
-	"github.com/google/gxui/math"
 	"github.com/google/gxui/mixins"
 )
 
-func CreateLabel(theme *Theme) gxui.Label {
-	l := &mixins.Label{}
-	l.Init(l, theme, theme.defaultFont, theme.LabelStyle.FontColor)
-	l.SetMargin(math.Spacing{L: 3, T: 3, R: 3, B: 3})
+func CreateLinearLayout(theme *Theme) gxui.LinearLayout {
+	l := &mixins.LinearLayout{}
+	l.Init(l, theme)
 	return l
 }
