+++ themes/basic/textbox.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package dark
+package basic
 
 import (
 	"github.com/google/gxui"
@@ -17,7 +17,7 @@ type TextBox struct {
 
 func CreateTextBox(theme *Theme) gxui.TextBox {
 	t := &TextBox{}
-	t.Init(t, theme.driver, theme, theme.defaultFont)
+	t.Init(t, theme.Driver(), theme, theme.DefaultFont())
 	t.SetTextColor(theme.TextBoxDefaultStyle.FontColor)
 	t.SetMargin(math.Spacing{L: 3, T: 3, R: 3, B: 3})
 	t.SetPadding(math.Spacing{L: 3, T: 3, R: 3, B: 3})
